+++ pontis-ui/src/components/Code/CodeLine.tsx
@@ -17,10 +17,11 @@ const CodeLine: React.FC<CodeLineProps> = ({ text, prefix }) => (
     className={classNames("relative px-10 text-opacity-70", codeStyles[prefix])}
   >
     <span className="absolute left-1 flex h-full items-center">
-      {prefix === "output" && (
+      {prefix === "output" ? (
         <ChevronRightIcon className="h-8 w-8 opacity-50" />
+      ) : (
+        <CheckIcon className="m-2 h-6 w-6 opacity-50" />
       )}
-      {prefix === "check" && <CheckIcon className="m-2 h-6 w-6 opacity-50" />}
     </span>
     <span className="py-2">{text}</span>
   </div>
