+++ pontis-ui/src/components/common/SectionContainer.tsx
@@ -10,9 +10,6 @@ interface ContainerProps {
   >;
 }
 
-// Alternative styles
-// px-6 pb-40 pt-12 sm:px-24 sm:pt-24 md:px-32 md:pt-32
-
 const SectionContainer: React.FC<ContainerProps> = ({
   className,
   children,
